ai/user_service/requirements.txt
pymongo 
bcrypt 
PyJWT 
fastapi uvicorn